Move Chat and User models with dependency chain documentation: implement chat conversation and message handling, user management functionalities, and document file dependencies for better clarity and organization.
