+++ backend/models/chat/chat.js
@@ -1,6 +1,6 @@
 import { v4 as uuidv4 } from 'uuid';
-import DbService from '../services/dbService.js';
-import logger from '../services/logger.js';
+import DbService from '../../services/dbService.js';
+import logger from '../../services/logger.js';
 
 /**
  * Create a new conversation in the database
