+++ backend/models/user/User.js
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
-import DbService from '../services/dbService.js';
+import DbService from '../../services/dbService.js';
 
 /**
  * User model for handling user-related database operations
